TestProject_Transom_2/requirements.txt
django==2.1

djangorestframework
django-filter
